requirements.txt
duckdb
pandas
requests
sodapy
python-dotenv
numpy
matplotlib
